+++ src/plugins/createPlugin.ts
@@ -9,15 +9,7 @@ import type { PersistedStore } from "../types"
 
 
 export function createPlugin(dbName?: string, cryptKey?: string, debug: boolean = false): PiniaPlugin {
-    const db = (!isEmpty(dbName) && dbName)
-        ? new Persister({ name: dbName, keyPath: 'storeName' })
-        : undefined
-
-    let crypt: Crypt | undefined
-
-    if (cryptKey) {
-        crypt = new Crypt(cryptKey)
-    }
+    const { db, crypt } = getEppsPluginOptions(dbName, cryptKey)
 
     if (typeof debug !== 'boolean') {
         debug = false
@@ -26,6 +18,27 @@ export function createPlugin(dbName?: string, cryptKey?: string, debug: boolean 
     const eppsPlugin = new EppsPlugin(db, crypt, debug)
 
     return eppsPlugin.plugin.bind(eppsPlugin)
+}
+
+function getEppsPluginOptions(dbName?: string, cryptKey?: string) {
+    let db: Persister | undefined
+    let crypt: Crypt | undefined
+
+    try {
+        if (window) {
+            if (!isEmpty(dbName) && dbName) {
+                db = new Persister({ name: dbName, keyPath: 'storeName' })
+            }
+
+            if (cryptKey) {
+                crypt = new Crypt(cryptKey)
+            }
+
+            return { db, crypt }
+        }
+    } catch (e) { }
+
+    return { db, crypt }
 }
 
 declare module 'pinia' {